grapes
Each ball is now shaded with a separate color, with intensity driven by band value.

The interface allows the user to pick two colors, the left endpoint and the right endpoint. The three balls are shaded, respectively, using the left color, an "interpolated" color between left and right, and the right color. Finally, the "brightness" of each color is driven by its respective band (which also controls the ball's size).