Compatible with triple balls.
James, you added rendering code to an old version. Watch out before you do this.

Don't we need to CREATE A SCENE before we can render? What is it rendering? How could it possibly render when the render routine never uses any of the scene parameters? I've added a 'gatherer()' call to the beginning of the render routine. I think this might be necessary.

Also, what is the render routine creating? What is an .iff file, and should all of them be going to the exact same filename? How do you change file destination when rendering? Finally, what is a .iff file, and can we get a quicktime movie here?